Allow labeled.json to contain results that refer to new examples.
This way a gui can inject additional examples for training that are not
a result from labeling an inferenced sample received from a scout.
At some point we would need to add a new RPC to send these labels with
the image content as 'LabeledExample' instead of referencing it through
the objectId as we do in 'SendLabel'.