[hawk-gui] Crop and resize new bootstrap examples.
Image.thumbnail maintains aspect ratio, so the example will end up
smaller than the tile size. We could pad the thumbnail with black
borders, but that would give us less to match on.
Instead, we crop the image to a square and then resize to the tile size.
This would lose pixel data from the edges, but keeps the center (focus?)
of the image at the best possible resolution.