[hawk-gui] Separate templates from missions, improve welcome page.
Adds an optional 'description' field to the mission_config.yml which is
expected to contain a markdown description for the selected template
that will be displayed on the welcome page.

Template missions are treated as read-only and the only provided action
is to create a new mission based on the template. To create a template,
create a directory where the name starts with '_' (underscore) and
inside of that directory create your mission_config.yml and a bootstrap
archive and if the user is not expected to modify the bootstrap
the optional pre-trained model file.