Fix weird dependency issue on exporting requirements.txt.
numpy > 1.24.4 dropped support for python 3.8 and so the poetry lock
file actually has two versions of numpy 'locked' depending on the python
version. But when we try to export to requirements.txt it complains
about the fact that the version we're using (1.26.4) doesn't support
all python versions the project supports.