Remove unused and under-documented a2s_new_model.
It was passing a path, binary data and version number. The version
number was consistently ignored and replaced by a new unique version.
If the path was pointing at a file on the scout (any file, anywhere) it
would get loaded in preference over the binary data. At least it wasn't
writing the binary data to the file because then anyone could overwrite
/etc/password, or ssh/authorized_keys...