grpcio-tools 1.72.x generates pb2 wrappers that do not work on python 3.8.
Poetry must have been using the lowest compatible version, while uv
tries to pull the highest version, or maybe it would have broken the
moment we had run poetry update on the lock file.

So either we need to build the wrappers on demand when we install/build
python wheels which may be tricky, or stick to an older grpcio-tools.